+++ lib/osm/model.rb
@@ -38,9 +38,33 @@ module Osm
     end
 
 
+    # Default to_i conversion is of id
     def to_i
-      id
+      id.to_i
     end
+
+    # Default compare based on id
+    def <=>(another)
+      return self.id <=> another.try(:id)
+    end
+
+    # Add other compare functions
+    def <(another)
+      send('<=>', another) < 0
+    end
+    def <=(another)
+      send('<=>', another) <= 0
+    end
+    def >(another)
+      send('<=>', another) > 0
+    end
+    def >=(another)
+      send('<=>', another) >= 0
+    end
+    def between?(min, max)
+      (send('<=>', min) > 0) && (send('<=>', max) < 0)
+    end
+
 
     # Get a list of attributes which have changed
     # @return Array[String] the names of attributes which have changed